mdbx: workaround MSVC C11-atomics and fix spilling internal accounting for `MDBX_AVOID_MSYNC=ON`.
2026-08-05 mdbx: update ChangeLog.
2026-08-05 mdbx: add comments and rename internal `sanitizer_probe_page_dangling()`.
2026-08-05 mdbx-cmake: use `CXX,C` for target-options as workaround cmake/visual-studio bugs.
2026-08-05 mdbx: fix copy&pase bugs in `sanitizer_kind_of_poison()` and `sanitizer_probe_page()`.
2026-08-05 mdbx-cmake: remove extra `/Oy` "omit frame pointers".
2026-08-05 mdbx-cmake: remove redundant `/utf-8`.
2026-08-05 mdbx-msvc: add compile-time guard for C11 atomics.
2026-08-05 mdbx-cmake: add `/experimental:c11atomics` for bad MSVC compiler to enable C11 atomics.
2026-08-05 mdbx-msvc: fixed MSVC warnings of implicit narrow-casting.
2026-08-05 mdbx-cmake: minor refine show compiler options.
2026-08-05 mdbx: fix copy&paste mistake in `key2float()`.
2026-08-04 mdbx: introduce `page_tmp_txnid_signature()` for clarity, checking and debugging.
2026-08-04 mdbx: push source page to the cursor's stash while splitting.
2026-08-04 mdbx: fix spilling/accounting for `MDBX_AVOID_MSYNC=ON`.
2026-08-03 mdbx: tracing of tmp-pages during splitting to avoid false-positive "dangling" issues.
2026-08-03 mdbx-tests: alter make-targets inside `stochastic` script.
2026-08-03 mdbx: drop minor unnecessary `if()` guard in `cursor_del()`.
2026-08-03 mdbx: refine/relax `txn_probe_dbi_cursors_stacks()`.
2026-08-03 mdbx-doc: minor clarify internal comments.

+++ mdbx_defrag.c
@@ -1,4 +1,4 @@
-/* This file is part of the libmdbx amalgamated source code (v0.14.2-508-g9abf372f at 2026-08-03T15:34:28+03:00).
+/* This file is part of the libmdbx amalgamated source code (v0.14.2-527-g6c0d0e15 at 2026-08-05T17:32:59+03:00).
  *
  * libmdbx (aka MDBX) is an extremely fast, compact, powerful, embeddedable, transactional key-value storage engine with
  * open-source code. MDBX has a specific set of properties and capabilities, focused on creating unique lightweight
@@ -176,7 +176,7 @@ static int defrag_notify(void *ctx, const MDBX_defrag_result_t *progress) {
     }
 
     if (ctx && tick) {
-      last_report_spenttime = progress->spent_time_dot16;
+      last_report_spenttime = (uint32_t)progress->spent_time_dot16;
       defrag_report_progress(progress, progress_dots++);
     }
     last_progress = *progress;
